+++ controller/handler_edge_ctrl/extend_enrollment_verify.go
@@ -17,9 +17,11 @@
 package handler_edge_ctrl
 
 import (
+	"bytes"
 	"fmt"
 	"github.com/michaelquigley/pfxlog"
 	"github.com/openziti/channel/v4"
+	nfpem "github.com/openziti/foundation/v2/pem"
 	"github.com/openziti/ziti/common/pb/edge_ctrl_pb"
 	"github.com/openziti/ziti/controller/env"
 	"google.golang.org/protobuf/proto"
@@ -82,7 +84,27 @@ func (h *extendEnrollmentVerifyHandler) HandleReceive(msg *channel.Message, ch c
 				return
 			}
 
-			if *edgeRouter.UnverifiedCertPem != verifyMsg.ClientCertPem {
+			submittedCerts := nfpem.PemStringToCertificates(verifyMsg.ClientCertPem)
+
+			if len(submittedCerts) == 0 {
+				h.respond(&edge_ctrl_pb.Error{
+					Code:    "INVALID_CLIENT_PEM",
+					Message: "request did not contain a parsable PEM certificate",
+				}, msg, ch)
+				return
+			}
+
+			targetCerts := nfpem.PemStringToCertificates(*edgeRouter.UnverifiedCertPem)
+
+			if len(targetCerts) == 0 {
+				h.respond(&edge_ctrl_pb.Error{
+					Code:    "INVALID_CLIENT_PEM",
+					Message: "internal error, target PEM did not parse to a certificate",
+				}, msg, ch)
+				return
+			}
+
+			if !bytes.Equal(targetCerts[0].Raw, submittedCerts[0].Raw) {
 				h.respond(&edge_ctrl_pb.Error{
 					Code:    "INVALID_CLIENT_PEM",
 					Message: "request did not contain a matching client pem",