+++ router/xgress_edge/certchecker_test.go
@@ -382,6 +382,8 @@ type SimpleTestIdentity struct {
 	setServerCertCalled bool
 }
 
+func (s *SimpleTestIdentity) IsCertSettable() error       { return nil }
+func (s *SimpleTestIdentity) IsServerCertSettable() error { return nil }
 func (s *SimpleTestIdentity) GetX509ActiveClientCertChain() []*x509.Certificate {
 	return nil
 }